[AIC-py][eval] bundle of small easy fixes
- freeze dataclasses
- wire up eval fn timeout further (there is a little more to go)
- move mock aiconfig runtime to separate file
-

see https://docs.google.com/document/d/15gx4vVmKVgI5bDHjSDd8xCQmTGcC8RrtLU0iCa32AIQ/edit
